# packages/docs/openapi/data-engine.yaml
openapi: 3.0.0
info:
    title: Data Engine API
    description: |
        API for managing structured data resources including domains, URLs, calls, tags, documents, tickets, and chats.

        This API provides comprehensive CRUD operations for managing data across multiple domains with support for filtering, pagination, and advanced search capabilities.
    version: 1.0.0
    contact:
        name: AnswerAgentAI Support
        url: https://theanswer.ai
        email: support@theanswer.ai
servers:
    - url: https://prod.studio.theanswer.ai/api/v1/data-engine
      description: Production server
    - url: http://localhost:4000/api/v1/data-engine
      description: Local development server
...
paths:
...
    /urls:
        post:
            summary: Create a new URL
            description: Create a new URL record
            operationId: createUrl
            tags:
                - urls
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UrlInput'
            responses:
                '201':
                    description: URL created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Url'
                '400':
                    description: Invalid input
                '401':
                    description: Unauthorized
        get:
            summary: List all URLs
            description: Retrieve URLs with filtering and pagination
            operationId: listUrls
            tags:
                - urls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - $ref: '#/components/parameters/SearchTermParam'
                - name: includeTags
                  in: query
                  schema:
                      type: string
                - name: excludeTags
                  in: query
                  schema:
                      type: string
                - name: hasAnalysis
                  in: query
                  schema:
                      type: string
                      enum: [all, has, not]
                - name: statusFilter
                  in: query
                  description: Array of HTTP status codes (JSON string)
                  schema:
                      type: string
                      example: '[200, 404]'
                - name: customFilters
                  in: query
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UrlListResponse'
                '401':
                    description: Unauthorized
        parameters:
            - $ref: '#/components/parameters/PageParam'
            - $ref: '#/components/parameters/PageSizeParam'
    /urls/{id}:
        get:
            summary: Get URL by ID
            operationId: getUrlById
            tags:
                - urls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Url'
                '404':
                    description: URL not found
        put:
            summary: Update URL
            operationId: updateUrl
            tags:
                - urls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UrlUpdate'
            responses:
                '200':
                    description: URL updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Url'
        delete:
            summary: Delete URL
            operationId: deleteUrl
            tags:
                - urls
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: URL deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
    # ==================== CALLS ====================
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: API Key
            description: |
                API key authentication. Include your API key in the Authorization header:
                `Authorization: Bearer YOUR_API_KEY`
    parameters:
        PageParam:
            name: page
            in: query
            description: Page number (0-indexed)
            schema:
                type: integer
                default: 0
                minimum: 0
        PageSizeParam:
            name: pageSize
            in: query
            description: Items per page
            schema:
                type: integer
                default: 10
                minimum: 1
                maximum: 100
        SearchTermParam:
            name: searchTerm
            in: query
            description: Search term for filtering results
            schema:
                type: string
        IdParam:
            name: id
            in: path
            required: true
            description: Resource ID (UUID)
            schema:
                type: string
                format: uuid
    schemas:
...
        Metadata:
            type: object
            description: Metadata object for tracking source system and user information
            properties:
                source_system:
                    type: string
                    example: theanswer
                source_organization_id:
                    type: string
                    format: uuid
                source_user_id:
                    type: string
                    format: uuid
                created_by:
                    type: string
                    format: email
                last_updated_by:
                    type: string
                    format: email
                last_updated_from:
                    type: string
                    example: theanswer
        DeleteResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: Resource deleted successfully
        PaginationInfo:
            type: object
            properties:
                total:
                    type: integer
                    example: 100
                page:
                    type: integer
                    example: 0
                pageSize:
                    type: integer
                    example: 10
        # Domain Schemas
...
        UrlInput:
            type: object
            required:
                - url
            properties:
                domain_id:
                    type: string
                    format: uuid
                domain_name:
                    type: string
                    example: example.com
                url:
                    type: string
                    format: uri
                    example: https://example.com/page
                http_status:
                    type: integer
                    example: 200
                status_text:
                    type: string
                    example: OK
                content_type:
                    type: string
                    example: text/html
                page_title:
                    type: string
                    example: Example Page
                meta_description:
                    type: string
                canonical_url:
                    type: string
                    format: uri
                custom_data:
                    type: object
                    additionalProperties: true
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Url:
            allOf:
                - $ref: '#/components/schemas/UrlInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        UrlUpdate:
            type: object
            properties:
                page_title:
                    type: string
                http_status:
                    type: integer
                custom_data:
                    type: object
                    additionalProperties: true
                metadata:
                    $ref: '#/components/schemas/Metadata'
        UrlListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      urls:
                          type: array
                          items:
                              $ref: '#/components/schemas/Url'
        # Call Schemas
...